--- src/test-fixtures/transform/01/output.json
@@ -1,342 +1,0 @@
-{
- "leader": "00000nab a22000005i 4500",
- "fields": [
-  {
-   "tag": "005",
-   "value": "00000000000000.0"
-  },
-  {
-   "tag": "008",
-   "value": "000101s        fi || || ||||||   ||fin c"
-  },
-  {
-   "tag": "041",
-   "ind1": " ",
-   "ind2": " ",
-   "subfields": [
-    {
-     "code": "a",
-     "value": "fin"
-    },
-    {
-     "code": "b",
-     "value": "fre"
-    },
-    {
-     "code": "b",
-     "value": "est"
-    },
-    {
-     "code": "b",
-     "value": "spa"
-    },
-    {
-     "code": "b",
-     "value": "smi"
-    },
-    {
-     "code": "b",
-     "value": "sme"
-    }
-   ]
-  },
-  {
-   "tag": "080",
-   "ind1": " ",
-   "ind2": " ",
-   "subfields": [
-    {
-     "code": "a",
-     "value": "021"
-    },
-    {
-     "code": "x",
-     "value": "2018"
-    }
-   ]
-  },
-  {
-   "tag": "084",
-   "ind1": " ",
-   "ind2": " ",
-   "subfields": [
-    {
-     "code": "a",
-     "value": "78.541"
-    },
-    {
-     "code": "2",
-     "value": "ykl"
-    }
-   ]
-  },
-  {
-   "tag": "084",
-   "ind1": " ",
-   "ind2": " ",
-   "subfields": [
-    {
-     "code": "a",
-     "value": "Eläimet "
-    },
-    {
-     "code": "2",
-     "value": "ykl"
-    }
-   ]
-  },
-  {
-   "tag": "336",
-   "ind1": " ",
-   "ind2": " ",
-   "subfields": [
-    {
-     "code": "a",
-     "value": "teksti"
-    },
-    {
-     "code": "b",
-     "value": "txt"
-    },
-    {
-     "code": "2",
-     "value": "rdacontent"
-    }
-   ]
-  },
-  {
-   "tag": "337",
-   "ind1": " ",
-   "ind2": " ",
-   "subfields": [
-    {
-     "code": "a",
-     "value": "käytettävissä ilman laitetta"
-    },
-    {
-     "code": "b",
-     "value": "n"
-    },
-    {
-     "code": "2",
-     "value": "rdamedia"
-    }
-   ]
-  },
-  {
-   "tag": "500",
-   "ind1": " ",
-   "ind2": " ",
-   "subfields": [
-    {
-     "code": "a",
-     "value": "Tämä on Yleinen huomautus"
-    }
-   ]
-  },
-  {
-   "tag": "520",
-   "ind1": " ",
-   "ind2": " ",
-   "subfields": [
-    {
-     "code": "a",
-     "value": ""
-    }
-   ]
-  },
-  {
-   "tag": "520",
-   "ind1": " ",
-   "ind2": " ",
-   "subfields": [
-    {
-     "code": "a",
-     "value": ""
-    }
-   ]
-  },
-  {
-   "tag": "520",
-   "ind1": " ",
-   "ind2": " ",
-   "subfields": [
-    {
-     "code": "a",
-     "value": ""
-    }
-   ]
-  },
-  {
-   "tag": "520",
-   "ind1": " ",
-   "ind2": " ",
-   "subfields": [
-    {
-     "code": "a",
-     "value": ""
-    }
-   ]
-  },
-  {
-   "tag": "567",
-   "ind1": " ",
-   "ind2": " ",
-   "subfields": [
-    {
-     "code": "a",
-     "value": "metodi ei tiedossa."
-    }
-   ]
-  },
-  {
-   "tag": "591",
-   "ind1": " ",
-   "ind2": " ",
-   "subfields": [
-    {
-     "code": "d",
-     "value": [
-      ""
-     ]
-    },
-    {
-     "code": "h",
-     "value": "Avaruustieteet ja tähtitiede"
-    },
-    {
-     "code": "i",
-     "value": "115"
-    },
-    {
-     "code": "h",
-     "value": "Geotieteet"
-    },
-    {
-     "code": "i",
-     "value": "1171"
-    },
-    {
-     "code": "5",
-     "value": "ARTO"
-    }
-   ]
-  },
-  {
-   "tag": "593",
-   "ind1": " ",
-   "ind2": " ",
-   "subfields": [
-    {
-     "code": "a",
-     "value": "todo"
-    },
-    {
-     "code": "5",
-     "value": "ARTO"
-    }
-   ]
-  },
-  {
-   "tag": "648",
-   "ind1": " ",
-   "ind2": "7",
-   "subfields": [
-    {
-     "code": "a",
-     "value": "1800-talet"
-    },
-    {
-     "code": "2",
-     "value": "yso/swe"
-    },
-    {
-     "code": "0",
-     "value": "http://www.yso.fi/onto/yso/p6180061899"
-    }
-   ]
-  },
-  {
-   "tag": "650",
-   "ind1": " ",
-   "ind2": "7",
-   "subfields": [
-    {
-     "code": "a",
-     "value": "talous"
-    },
-    {
-     "code": "2",
-     "value": "yso/fin"
-    },
-    {
-     "code": "0",
-     "value": "http://www.yso.fi/onto/yso/p2555"
-    }
-   ]
-  },
-  {
-   "tag": "650",
-   "ind1": " ",
-   "ind2": "7",
-   "subfields": [
-    {
-     "code": "a",
-     "value": "tekniikat"
-    },
-    {
-     "code": "2",
-     "value": "yso/fin"
-    },
-    {
-     "code": "0",
-     "value": "http://www.yso.fi/onto/yso/p38430"
-    }
-   ]
-  },
-  {
-   "tag": "773",
-   "ind1": "0",
-   "ind2": " ",
-   "subfields": [
-    {
-     "code": "7",
-     "value": "nnas"
-    },
-    {
-     "code": "w",
-     "value": [
-      "(FI-MELINDA)000434823"
-     ]
-    },
-    {
-     "code": "t",
-     "value": "Basso : urbaani musiikki & kulttuuri. -"
-    },
-    {
-     "code": "d",
-     "value": "Helsinki : Basso Media, 2006-2012."
-    },
-    {
-     "code": "x",
-     "value": "1796-3753. -"
-    },
-    {
-     "code": "g",
-     "value": " () : , sivut "
-    }
-   ]
-  },
-  {
-   "tag": "960",
-   "ind1": " ",
-   "ind2": " ",
-   "subfields": [
-    {
-     "code": "a",
-     "value": "ARTO"
-    }
-   ]
-  }
- ]
-}
