--- src/test-fixtures/transform/02/info.txt
@@ -1,10 +1,0 @@
-
-02:
-
-generate0xxFields.js
----------------------
-
-functions used and tested here:
-041: artikkelin & tiivistelmäm kieli (spa, saame etc)
-080: Lisäkentät a & x
-084: Muu luokitus (084 $a) & Luokituksen lähde (084 $2):
