--- src/test-fixtures/transform/03/info.txt
@@ -1,10 +1,0 @@
-03:
-
-generate1xxFields:
-------------------
-
-functions used and tested here:
-
-generatef100sf110sf700sf710s: Lisää tekijä
-
-cumulative, this includes previous inputs too.
