+++ test-fixtures/generatef041/01/input.json
@@ -1,0 +1,5 @@
+{
+    "articleLanguage": "fin",
+    "abstractLanguages": [ "sme", "fre", "est", "esp", "smn", "smi" ]
+}
+
