+++ test-fixtures/generatef041/01/output.json
@@ -1,0 +1,37 @@
+[
+  {
+   "tag": "041",
+   "ind1": " ",
+   "ind2": " ",
+   "subfields": [
+    {
+     "code": "a",
+     "value": "fin"
+    },
+    {
+     "code": "b",
+     "value": "fre"
+    },
+    {
+     "code": "b",
+     "value": "est"
+    },
+    {
+     "code": "b",
+     "value": "spa"
+    },
+    {
+     "code": "b",
+     "value": "smi"
+    },
+    {
+     "code": "b",
+     "value": "sme"
+    },
+    {
+     "code": "b",
+     "value": "smn"
+    }
+   ]
+  }
+]
