+++ test-fixtures/generatef080/01/metadata.json
@@ -1,0 +1,8 @@
+{
+  "description": "generatef080; field 080",
+  "disabled": false,
+  "succesRecords": 2,
+  "failedRecords": 0,
+  "expectedError": false,
+  "expectedErrorStatus": "200"
+}
