+++ test-fixtures/generatef084/01/metadata.json
@@ -1,0 +1,8 @@
+{
+  "description": "generatef084; field 084",
+  "disabled": false,
+  "succesRecords": 2,
+  "failedRecords": 0,
+  "expectedError": false,
+  "expectedErrorStatus": "200"
+}
