+++ test-fixtures/generatef084/01/output.json
@@ -1,0 +1,33 @@
+  [
+  {
+   "tag": "084",
+   "ind1": " ",
+   "ind2": " ",
+   "subfields": [
+    {
+     "code": "a",
+     "value": "78.541"
+    },
+    {
+     "code": "2",
+     "value": "ykl"
+    }
+   ]
+  },
+  {
+   "tag": "084",
+   "ind1": " ",
+   "ind2": " ",
+   "subfields": [
+    {
+     "code": "a",
+     "value": "Eläimet "
+    },
+    {
+     "code": "2",
+     "value": "ykl"
+    }
+   ]
+  }
+]
+
