+++ test-fixtures/generatef084/02/input.json
@@ -528,7 +528,7 @@
    "two084": "ykl"
   },
   {
-   "key": 3,
+   "key": 2,
    "a084": "Eläimet ",
    "two084": "ykl"
   }
