+++ test-fixtures/generatef084/02/metadata.json
@@ -1,0 +1,5 @@
+{
+  "description": "This test expected to fail: 422",
+  "expectToFail": true,
+  "expectedFailStatus": 422
+}
