+++ test-fixtures/generatef084/03/input.json
@@ -1,0 +1,1 @@
+{}
